9624: Move ConditionModel.js -> ConditionFormField/model.js
